refactor: add FluentCommentsParser class
